Refactor package structure for better project organization
Updated package paths to follow a clearer hierarchy under `com.github.ansgrb.rockpaperscissorsapi`. Moved shared models to a separate shared module for better modularization and reusability. Adjusted imports and references across the project to reflect these changes.
