+++ server/src/main/kotlin/com/github/ansgrb/rockpaperscissors/server/Application.kt
@@ -1,9 +1,10 @@
-package dev.ansgrb
+package com.github.ansgrb.rockpaperscissors.server
 
 import io.ktor.server.application.*
+import io.ktor.server.netty.EngineMain
 
 fun main(args: Array<String>) {
-	io.ktor.server.netty.EngineMain.main(args)
+	EngineMain.main(args)
 }
 
 fun Application.module() {
