+++ server/src/main/kotlin/com/github/ansgrb/rockpaperscissors/server/Cors.kt
@@ -1,11 +1,10 @@
-package dev.ansgrb
+package com.github.ansgrb.rockpaperscissors.server
 
 import io.ktor.http.HttpMethod
 import io.ktor.server.plugins.cors.routing.CORS
 import io.ktor.http.HttpHeaders
 import io.ktor.server.application.Application
 import io.ktor.server.application.install
-import io.ktor.server.plugins.cors.routing.CORS
 
 fun Application.configureCORS() {
 	install(CORS) {
