+++ server/src/main/kotlin/com/github/ansgrb/rockpaperscissors/server/Serialization.kt
@@ -1,14 +1,10 @@
-package dev.ansgrb
+package com.github.ansgrb.rockpaperscissors.server
 
 import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.contentnegotiation.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import io.ktor.server.websocket.*
-import io.ktor.websocket.*
-import java.time.Duration
-import kotlin.time.Duration.Companion.seconds
 
 fun Application.configureSerialization() {
 	install(ContentNegotiation) {
