+++ shared/build.gradle.kts
@@ -5,5 +5,4 @@ plugins {
 
 dependencies {
 	implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.0")
-	// Add other shared dependencies
 }
